Allow options to override default values
options were being ignored and were unable to override default layer
settings such as maxZoom.

Note: it would also be possible to merge wmsParams in to options so that they too can be overridden by user options. However, in this case, Leaflet would change width and height settings for retina screens and I'm unable to test if this would work so I left it alone.